@@ -5,11 +5,11 @@ end
 source "https://rubygems.org"
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby "3.0.2"
+ruby "3.1.3"
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 if next?
-  gem "rails", github: "rails/rails", branch: "main"
+  gem "rails", "~> 7.0.0"
 else
   gem "rails", "~> 6.1.0"
 end
