@@ -1,3 +1,3 @@
-class AnalyzedFile < ActiveRecord::Base
+class AnalyzedFile < ApplicationRecord
   belongs_to :report
 end
