@@ -1,5 +1,5 @@
-class Project < ActiveRecord::Base
-  has_many :reports
+class Project < ApplicationRecord
+  has_many :reports, dependent: :destroy
 
   validates :permalink, uniqueness: true
 end
