@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_11_06_175125) do
+ActiveRecord::Schema.define(version: 2023_03_20_185535) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -45,6 +45,7 @@ ActiveRecord::Schema.define(version: 2021_11_06_175125) do
     t.string "slug"
     t.bigint "project_id"
     t.index ["project_id"], name: "index_reports_on_project_id"
+    t.index ["slug"], name: "index_reports_on_slug", unique: true
   end
 
   add_foreign_key "reports", "projects"
